refactor: Lotto Numbers 원시타입 객체 포장 -> TestCode 에는 영향이 없게.
